+++ Source/SwiftLintBuiltInRules/Models/BuiltInRules.swift
@@ -126,6 +126,7 @@ public let builtInRules: [any Rule.Type] = [
     NSObjectPreferIsEqualRule.self,
     NestingRule.self,
     NimbleOperatorRule.self,
+    NoEmptyBlockRule.self,
     NoExtensionAccessModifierRule.self,
     NoFallthroughOnlyRule.self,
     NoGroupingExtensionRule.self,