+++ CHANGELOG.md
@@ -30,6 +30,11 @@
   [Martin Redington](https://github.com/mildm8nnered)
   [#5552](https://github.com/realm/SwiftLint/issues/5552)
 
+* Add `no_empty_block` default rule to validate that code blocks are not empty.
+  They should at least contain a comment.  
+    [Ueeek](https://github.com/Ueeek)
+    [#5615](https://github.com/realm/SwiftLint/issues/5615)
+
 #### Bug Fixes
 
 * Fix a few false positives and negatives by updating the parser to support